moved linked list functions to card.c
also added ascii values for card values
